+++ CHANGELOG.md
@@ -65,6 +65,7 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#85](https://github.com/ARGOeu/argo-mon-status-api/pull/85) ARGO-5361 Fetch Tenant's Reports
 - [#91](https://github.com/ARGOeu/argo-mon-status-api/pull/91) ARGO-5369 Exclude CHECK_READINESS from jobs list in status
 - [#93](https://github.com/ARGOeu/argo-mon-status-api/pull/93) ARGO-5373 Reports order by active first
+- [#94](https://github.com/ARGOeu/argo-mon-status-api/pull/94/) ARGO-5374 List tenant status pages based on tenant role
 
 ### Fix